Changed back the default behavior for the overall set of downloads
IMHO for first-time users, it's going to be confusing if the final
action is to bag and archive the archived bags it produces.  Also, I
don't think it's often not desirable to create a single archive of
everything, especially in the case where the results are meant to be
put in a cloud-based storage facility.  In that case, any retrieval
request would result in having to retrieve the entire archive, which
can be costly for large archive files.  I think many users are more
likely to use, at most, an uncompressed bag of all the downloads, so
that the individual zip'ed bags are individually retrievable from
whatever archive service they're using.