Significant changes to support creating overall bag
`eprints2bags` is now able to create a top-level bag containing the
archived bags it creates.  This top-level bag itself can also be put
into a single-file archive if desired.  This behavior is controlled by
the new option `-e` in combination with the (renamed) option `-b` and
the new option `-t`.

The implementation required refactoring some code, and also changing
around some of the command-line option letters in order to try to find
a combination that was reasonably readable and intuitive.