Don't use multiple processes for small bags
Turns out that if you hand bagit.make_bag() a number of processes > 1,
it will always create a multiprocess pool of that size, and creating
that pool is slow.  The new code checks that the number of items in a
bag directory is more than the number of processes requested before it
uses multiple processes.