Fix: setuptools doesn't deal with multiple authors in `author` field
As others have noted (e.g., https://stackoverflow.com/q/9999829/743730),
it turns out that setuptools doesn't handle multiple authors in the
`author` and `author_email` fields.  Although I don't like it, I'm
reducing the values to a single name for the main author (me), which
works in this case because all of the code in the Python version of
eprints2bags is truly my own.  The README and other documentation
mention Betsy Coles' contribution, and thus it shows up elsewhere on
the PyPi page for this package.  Thus, I think having only me in the
`author` value is a reasonable compromise for this particular software
package and does not constitute erasure of Betsy's contributions.